requirements.txt
asgiref==3.2.10
boto3==1.14.56
botocore==1.17.56
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.1
django-ckeditor==6.0.0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-storages==1.10
djangorestframework==3.11.1
docutils==0.15.2
google-api-core==1.22.2
google-auth==1.21.1
google-cloud-core==1.4.1
google-cloud-storage==1.31.0
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
Markdown==3.2.2
Pillow==7.2.0
protobuf==3.13.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
rsa==4.6
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
